+++ src/docs/previews/dialog/nested/css/index.svelte
@@ -15,7 +15,9 @@
 			portalled,
 		},
 		states: { open },
-	} = createDialog();
+	} = createDialog({
+		forceVisible: true,
+	});
 
 	const {
 		elements: {
@@ -28,7 +30,9 @@
 			portalled: portalledNested,
 		},
 		states: { open: openNested },
-	} = createDialog();
+	} = createDialog({
+		forceVisible: true,
+	});
 </script>
 
 <button use:melt={$trigger} class="trigger"> Open Dialog </button>